+++ pom.xml
@@ -8,7 +8,7 @@
     <parent>
         <groupId>com.vaadin</groupId>
         <artifactId>vaadin-parent</artifactId>
-        <version>2.0.5</version>
+        <version>2.1.0</version>
     </parent>
 
     <artifactId>sso-kit-hilla</artifactId>
@@ -24,8 +24,8 @@
         <maven.compiler.target>11</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <hilla.version>1.3-SNAPSHOT</hilla.version>
-        <flow.version>23.3-SNAPSHOT</flow.version>
+        <hilla.version>1.3.1</hilla.version>
+        <flow.version>23.3.1</flow.version>
         <spring.boot.version>2.7.4</spring.boot.version>
         <maven.jar.version>3.3.0</maven.jar.version>
         <maven.source.version>3.2.1</maven.source.version>
