+++ sso-kit-hilla-demo/src/main/java/com/example/application/endpoints/helloworld/HelloWorldEndpoint.java
@@ -2,10 +2,9 @@ package com.example.application.endpoints.helloworld;
 
 import javax.annotation.security.PermitAll;
 
-import com.vaadin.sso.hilla.starter.UserEndpoint;
-
 import dev.hilla.Endpoint;
 import dev.hilla.Nonnull;
+import dev.hilla.sso.endpoint.AuthEndpoint;
 
 @Endpoint
 @PermitAll
@@ -15,7 +14,7 @@ public class HelloWorldEndpoint {
     public String sayHello(@Nonnull String name) {
         if (name.isEmpty()) {
             return "Hello "
-                    + new UserEndpoint().getAuthenticatedUser().get().getName();
+                    + new AuthEndpoint().getAuthenticatedUser().get().getName();
         } else {
             return "Hello " + name;
         }
