+++ sso-kit-hilla-starter/src/main/java/dev/hilla/sso/starter/Configuration.java
@@ -1,0 +1,4 @@
+package dev.hilla.sso.starter;
+
+public class Configuration {
+}
