+++ sso-kit-hilla-starter/src/main/java/dev/hilla/sso/starter/SingleSignOnProperties.java
@@ -1,4 +1,4 @@
-package com.vaadin.sso.hilla.starter;
+package dev.hilla.sso.starter;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
