+++ sso-kit-hilla-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-com.vaadin.sso.hilla.starter.SingleSignOnProperties
+dev.hilla.sso.starter.SingleSignOnProperties
